# app/Http/Virtual/GraphUpdateRequest.php
<?php

namespace App\Http\Virtual;

/**
 * @OA\Schema(
 *      title="Update Graph request",
 *      description="Update Graph request body data",
 *      type="object",
 *      required={"name"}
 * )
 */
class GraphUpdateRequest
{

}